#346: extract shared isOwnAnnounce, sync discovery.md, lift InMemoryFingerprintPersistence
- SelfSuppression.isOwnAnnounce(peerFingerprint, ownFingerprint) shared
  between MdnsDiscoveryJmdns and bonjour/BonjourState.
- discovery.md §Identity and self-suppression now distinguishes per-install
  (apps) from per-process (CLI), with a cross-link to the knowledge note.
- DeviceIdentityStoreTest uses commonTest InMemoryFingerprintPersistence
  instead of an inline duplicate.
