+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/identity/SelfSuppression.kt
@@ -1,0 +1,4 @@
+package com.tubetoast.tether.identity
+
+fun isOwnAnnounce(peerFingerprint: String?, ownFingerprint: String): Boolean =
+    peerFingerprint != null && peerFingerprint == ownFingerprint
