+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/discovery/MdnsDiscoveryJmdns.kt
@@ -1,6 +1,7 @@
 package com.tubetoast.tether.discovery
 
 import com.tubetoast.tether.identity.DeviceIdentityStore
+import com.tubetoast.tether.identity.isOwnAnnounce
 import com.tubetoast.tether.protocol.Device
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
@@ -220,7 +221,7 @@ internal class MdnsDiscoveryJmdns(
     }
 
     internal fun isOwnAnnounce(peerFingerprint: String?): Boolean =
-        peerFingerprint != null && peerFingerprint == fingerprint
+        isOwnAnnounce(peerFingerprint, fingerprint)
 
     /** JmDNS resolves A/AAAA in stages; first callback may carry only IPv6. Returns `null` to wait for next event. */
     private fun resolveIPv4(info: ServiceInfo, serviceName: String): String? {
