+++ docs/engineering/discovery.md
@@ -157,7 +157,9 @@ Any combination of failure modes Layers 1–3 cannot overcome, the user works ar
 
 The `fingerprint` field in `/hello` and `PeerAnnouncement` carries a stable device identity. Its target is the EC P-256 public key fingerprint produced by [Pairing (#11)](https://github.com/khmelevartem/tether/issues/11) — the same identity used by [channel encryption](../product/security.md#channel-encryption) and pairing.
 
-Until pairing identity lands, the field carries a per-install random opaque string sufficient for self-suppression but not for trust. No code path treats it as authentication; trust gating remains pairing. The interim string is regenerated on app reinstall, matching the user-visible "reinstall produces a new identity" behaviour of pairing.
+Until pairing identity lands, the field carries a random opaque string sufficient for self-suppression but not for trust. No code path treats it as authentication; trust gating remains pairing.
+
+For app installs (Android, iOS, Desktop UI) the string is persisted via DataStore and regenerated on reinstall, matching the user-visible "reinstall produces a new identity" behaviour of pairing. For the CLI (`tether-cli`) it is per-process ephemeral — never persisted, so two CLI processes on the same host get distinct identities. See [`docs/knowledge/cli-multi-instance.md`](../knowledge/cli-multi-instance.md).
 
 This dependency is reflected as a `TODO` in the discovery implementation referencing [#11](https://github.com/khmelevartem/tether/issues/11). It is not reflected in this document beyond this section, because the *contract* of the `fingerprint` field does not change between interim and final identity — only the source.
